requirements.txt
cat > requirements.txt <<EOF
torch
torchvision
scikit-learn
tqdm
Pillow
EOF
